chore: reorganize test files into logical folders (#114)
