+++ .github/workflows/main.yml
@@ -1,9 +1,10 @@
 name: Main Branch Checks
 
 on:
+    push:
+        branches: [main]
     pull_request:
         types: [closed]
-    push:
         branches: [main]
 
 jobs:
