+++ package.json
@@ -6,13 +6,15 @@
 		"build": "tsc",
 		"start": "node dist/bunkbot.js",
 		"dev": "npm run lint:fix && npm run build && npm run test",
-		"test": "jest --config=jest.config.js",
+		"test": "jest",
 		"tests": "jest --config=jest.config.js",
+		"test:coverage": "jest --coverage",
 		"test:watch": "jest --watch",
 		"lint": "eslint 'src/**/*.{ts,js}' '*.{ts,js}'",
 		"lint:fix": "eslint 'src/**/*.{ts,js}' '*.{ts,js}' --fix",
 		"format": "prettier --write \"src/**/*.ts\"",
-		"type-check": "tsc --noEmit"
+		"type-check": "tsc --noEmit",
+		"test:music": "jest src/tests/starbunk/commands/musicCommands.test.ts"
 	},
 	"engines": {
 		"node": ">=16.0.0"
@@ -65,17 +67,5 @@
 		"tsconfig-paths": "^4.2.0",
 		"tslib": "^2.6.2",
 		"typescript": "^5.0.0"
-	},
-	"jest": {
-		"testEnvironment": "node",
-		"testMatch": [
-			"<rootDir>/src/tests/**/*.test.ts"
-		],
-		"moduleNameMapper": {
-			"^@/(.*)$": "<rootDir>/src/$1"
-		},
-		"setupFilesAfterEnv": [
-			"<rootDir>/src/tests/setup.ts"
-		]
 	}
 }
