+++ src/__tests__/mocks/openaiMocks.ts
@@ -1,0 +1,11 @@
+export const mockOpenAI = {
+	chat: {
+		completions: {
+			create: jest.fn()
+		}
+	}
+};
+
+jest.mock('@/openai/openaiClient', () => ({
+	openai: mockOpenAI
+}));
